+++ src/page/manager.rs
@@ -3,7 +3,7 @@ use crate::page::PageId;
 #[cfg(feature = "buffer_pool_backend")]
 pub(super) mod buffer_pool;
 #[cfg(feature = "buffer_pool_backend")]
-pub(super) mod clock_replacer;
+pub(super) mod frame;
 #[cfg(feature = "mmap_backend")]
 pub(super) mod mmap;
 pub(super) mod options;
