+++ src/page/manager/buffer_pool.rs
@@ -1,7 +1,6 @@
 use crossbeam_channel::{Receiver, Sender};
 use fxhash::FxBuildHasher;
 use io_uring::{opcode, types, IoUring};
-use proptest::num;
 use std::{
     ffi::CString,
     fs::File,
@@ -23,110 +22,12 @@ use parking_lot::{Mutex, RwLock};
 
 use crate::{
     page::{
-        manager::clock_replacer::ClockReplacer,
+        manager::frame::{Frame, FrameId, Frames},
         state::{PageState, RawPageState},
         Page, PageError, PageId, PageManagerOptions, PageMut,
     },
     snapshot::SnapshotId,
 };
-
-#[derive(Debug)]
-struct Frame {
-    ptr: *mut [u8; Page::SIZE],
-    page_id: AtomicU32, // 0 means None, otherwise it's the page_id
-    state: AtomicU32,   // Bit 0: pin (If true, this frame cannot be evicted)
-                        // Bit 1: ref_bit (Second Chance bit)
-}
-
-impl Clone for Frame {
-    fn clone(&self) -> Self {
-        Frame {
-            ptr: self.ptr,
-            page_id: AtomicU32::new(self.page_id.load(Ordering::SeqCst)),
-            state: AtomicU32::new(self.state.load(Ordering::SeqCst)),
-        }
-    }
-}
-
-impl Frame {
-    // Unpin the frame if the page is not dirty.
-    // If the page is in dirty state, returns false.
-    fn unpin(&self) -> bool {
-        let is_dirty = unsafe { PageState::is_dirty(self.ptr.cast()) };
-
-        if is_dirty {
-            return false;
-        }
-        // Clear pin bit (bit 0), keep ref_bit (bit 1) unchanged
-        self.state.fetch_and(0b10, Ordering::Release);
-        true
-    }
-
-    #[inline(always)]
-    fn pin(&self) {
-        // Set both pin (bit 0) and ref_bit (bit 1) to true
-        self.state.store(0b11, Ordering::Release);
-    }
-}
-
-// SAFETY: Frame contains a pointer to heap-allocated memory that we own exclusively.
-// The memory is allocated via Box and we manage its lifetime, so it's safe to send
-// between threads.
-unsafe impl Send for Frame {}
-unsafe impl Sync for Frame {}
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
-pub(crate) struct FrameId(u32);
-
-impl FrameId {
-    #[inline(always)]
-    pub(crate) const fn as_usize(&self) -> usize {
-        self.0 as usize
-    }
-
-    #[inline]
-    pub(crate) const fn from_usize(value: usize) -> Self {
-        FrameId(value as u32)
-    }
-}
-
-struct Frames(Vec<Frame>);
-impl Frames {
-    fn allocate(num_frames: usize) -> Self {
-        let mut frames = Vec::with_capacity(num_frames);
-        (0..num_frames).into_iter().for_each(|_| {
-            let boxed_array = Box::new([0; Page::SIZE]);
-            let ptr = Box::into_raw(boxed_array);
-            frames.push(Frame { ptr, page_id: AtomicU32::new(0), state: AtomicU32::new(0) });
-        });
-        Self(frames)
-    }
-
-    #[inline(always)]
-    fn get(&self, frame_id: FrameId) -> &Frame {
-        &self.0[frame_id.0 as usize]
-    }
-
-    // Unpin the frame if the occupied page is not dirty.
-    // If the page is in dirty state, returns false.
-    fn unpin(&self, frame_id: FrameId) -> bool {
-        let frame = self.get(frame_id);
-        frame.unpin()
-    }
-
-    // Find a frame to be evicted, also pin that frame and running cleanup F function.
-    fn victim_and_pin<F>(&self, cleanup: F) -> Option<(FrameId, &Frame)>
-    where
-        F: FnOnce(FrameId),
-    {
-        todo!()
-    }
-
-    #[cfg(test)]
-    fn count_pinned(&self) -> usize {
-        self.0.iter().filter(|i| (i.state.load(Ordering::SeqCst) & 0b01) != 0).count()
-    }
-}
 
 pub(crate) type FxMap<K, V> = DashMap<K, V, FxBuildHasher>;
 pub(crate) type FxSet<K> = DashSet<K, FxBuildHasher>;
@@ -136,8 +37,6 @@ pub struct PageManager {
                                          * indexed by page id with fix num_frames size */
     frames: Arc<Frames>, /* list of frames that hold pages' data, indexed by frame id with
                           * fix num_frames size */
-    frame_lock_replacer_hand: Mutex<usize>, /* CLOCK cache replacer hand */
-    // replacer: Arc<ClockReplacer>,
     loading_pages: FxSet<PageId>,
     page_count: AtomicU32,
     num_frames: u32,
@@ -218,7 +117,6 @@ impl PageManager {
             file: RwLock::new(file),
             file_len,
             frames: Arc::new(frames),
-            frame_lock_replacer_hand: Mutex::new(0),
             page_table,
             updated_pages: Arc::new(DashMap::with_hasher(FxBuildHasher::default())),
             new_pages: Mutex::new(Vec::new()),
